Add modifications to work with Unity 2018.3.10f1
(See README.md for the background and a general explanation of
the changes.)

Modify source code to make it compatible with the .NET/C#
version used by Unity 2018.3.10f1. This version of the library
is meant to be used directly as source code, rather than a pre-
compiled package from NuGet. Generally speaking, this means
the code had to be adapted to C# 7.5, instead of C# 8.

All changes are applied to the library (Figgle folder) only.

The Figgle.csproj inside of the Figgle folder was deleted to
simplify the copying of the library in Unity.

The banner at the top of README.md was modified slightly with
a more compact version of the font, to avoid GitHub from
displaying it inside of a scroll view.